Improved the assertion monitoring framework
o Removed the type Assert from Control.Precondition
o Added type Pre of kind Constraint

functions with preconditions used to have type:

f :: Assert -> a -> b

and would be called "f assert x" so that the assertion failure would give the line number of the call to f instead of the definition to f. Now f has type:

f :: Pre => a -> b

and we call it as "f x". Pre is part of the context of the type, not the type of an argument. It has the following definition:

type Pre = (?loc :: CallStack)

This makes use of implicit named parameters. In particular, f :: (?loc :: CallStack) => a -> b is a function that takes an explicit value of type a and an implict value of type CallStack, with one reference to the line number of the call to f. In this context, f has Pre => in its type because it is a partial function. A precondition failure will give the location of the calling code.