Add 'directory < 1.2.6.0' constraint wherever MissingH is a dependency
The directory package introduced a `System.Directory.isSymbolicLink` in
version 1.2.6.0 which clashes with `System.Posix.Files.Common.isSymbolicLink`
from the unix package. The ambiguous import is in src/System/IO/HVFS.hs
in the MissingH package, the development of which seems to have stopped.
So, we're restricting directory to < 1.2.6.0 until the developer of
MissingH fixes the ambiguity or until we swap out MissingH.