[1/4] Swap out ConfigFile for yaml
Config file's name changed from z3_config.conf to z3_config.yml.

TODO: there's a slight issue with the yaml library treating all
of Z3Config's fields as String, in that it will require the numbers
to be explicitly quoted to be considered as string literals; and
otherwise it will ignore them. We'll try to fix this by generalizing
the Document type class and the functions using it.